requirements.txt
PyPDF2==3.0.1
aiohttp==3.11.13
asyncio==3.4.3
google-cloud-aiplatform
google-cloud-storage
psycopg2-binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
vertexai==1.71.1
